// src/spider/jili/proto/general/astarte2_196.ts
export interface Request {
    /**
     * @generated from protobuf field: serverProto.AckType ack = 1
     */
    ack: AckType;
    /**
     * @generated from protobuf field: bytes req = 2
     */
    req: Uint8Array;
}
// ...
export interface InfoReq {
    /**
     * @generated from protobuf field: string os = 1
     */
    os: string;
    /**
     * @generated from protobuf field: string language = 2
     */
    language: string;
    /**
     * @generated from protobuf field: serverProto.Browser browser = 3
     */
    browser?: Browser;
    /**
     * @generated from protobuf field: string version = 4
     */
    version: string;
    /**
     * @generated from protobuf field: string model = 5
     */
    model: string;
}
/**
 * 定义 Browser 消息类型，用于表示浏览器的相关信息
 *
 * @generated from protobuf message serverProto.Browser
 */
export interface Browser {
    /**
     * 浏览器的类型，如 Chrome、Firefox 等
     *
     * @generated from protobuf field: string type = 1
     */
    type: string;
    /**
     * 浏览器的版本号
     *
     * @generated from protobuf field: string version = 2
     */
    version: string;
    /**
     * 浏览器的语言设置
     *
     * @generated from protobuf field: string language = 3
     */
    language: string;
    /**
     * 浏览器窗口的宽度
     *
     * @generated from protobuf field: uint32 width = 4
     */
    width: number;
    /**
     * 浏览器窗口的高度
     *
     * @generated from protobuf field: uint32 height = 5
     */
    height: number;
    /**
     * 浏览器的比例
     *
     * @generated from protobuf field: float ratio = 6
     */
    ratio: number;
}
/**
 * 定义 GameInfoAck 消息类型，用于表示游戏信息的确认消息
 *
 * @generated from protobuf message serverProto.GameInfoAck
 */
export interface GameInfoAck {
    /**
     * 钱包信息列表，每个元素是一个 WalletInfo 消息
     *
     * @generated from protobuf field: repeated serverProto.WalletInfo WalletInfo = 1
     */
    walletInfo: WalletInfo[];
    /**
     * 偏好轮次显示信息，是一个 PreferRoundShow 消息
     *
     * @generated from protobuf field: serverProto.PreferRoundShow Prefer = 2
     */
    prefer?: PreferRoundShow;
    /**
     * 免费旋转剩余次数
     *
     * @generated from protobuf field: int32 FreeSpinRemain = 3
     */
    freeSpinRemain: number;
    /**
     * 额外信息，以字节数组形式存储
     *
     * @generated from protobuf field: bytes extraInfo = 5
     */
    extraInfo: Uint8Array;
    /**
     * 游戏的理论回报率
     *
     * @generated from protobuf field: double rtp = 6
     */
    rtp: number;
    /**
     * 解锁大奖的投注金额列表
     *
     * @generated from protobuf field: repeated double jpUnlockBet = 7
     */
    jpUnlockBet: number[];
    /**
     * 商城信息，是一个 MallInfo 消息
     *
     * @generated from protobuf field: serverProto.MallInfo mall = 8
     */
    mall?: MallInfo;
    /**
     * API 类型
     *
     * @generated from protobuf field: int32 ApiType = 9
     */
    apiType: number;
    /**
     * 免费旋转的投注金额
     *
     * @generated from protobuf field: double FreeSpinBet = 10
     */
    freeSpinBet: number;
    /**
     * 游戏商城信息，是一个 GameMallInfo 消息
     *
     * @generated from protobuf field: serverProto.GameMallInfo gameMall = 12
     */
    gameMall?: GameMallInfo;
    /**
     * 最大赔率
     *
     * @generated from protobuf field: double MaxOdd = 13
     */
    maxOdd: number;
    /**
     * 免费旋转列表信息，是一个 freeSpinList 消息
     *
     * @generated from protobuf field: serverProto.freeSpinList freeSpin = 14
     */
    freeSpin?: freeSpinList;
    /**
     * 免费旋转的类型
     *
     * @generated from protobuf field: int32 freeSpinType = 15
     */
    freeSpinType: number;
    /**
     * 锦标赛信息，是一个 Tournament 消息
     *
     * @generated from protobuf field: serverProto.Tournament tournament = 16
     */
    tournament?: Tournament;
}
// ...
export interface PreferRoundShow {
    /**
     * 基础轮次相关数值，使用 double 类型存储
     *
     * @generated from protobuf field: double BaseRound = 1
     */
    baseRound: number;
    /**
     * Sigma 轮次相关数值，使用 double 类型存储
     *
     * @generated from protobuf field: double SigmaRound = 2
     */
    sigmaRound: number;
    /**
     * 变异系数相关数值，使用 double 类型存储
     *
     * @generated from protobuf field: double CV = 3
     */
    cV: number;
    /**
     * 命中率相关数值，使用 double 类型存储
     *
     * @generated from protobuf field: double HR = 4
     */
    hR: number;
    /**
     * 最大连赢轮次相关数值，使用 double 类型存储
     *
     * @generated from protobuf field: double MGR = 5
     */
    mGR: number;
    /**
     * 范围，使用 int32 类型存储
     *
     * @generated from protobuf field: int32 Range = 6
     */
    range: number;
    /**
     * 展示标志，使用 int32 类型存储
     *
     * @generated from protobuf field: int32 Show = 7
     */
    show: number;
}
// ...
export interface SpinReq {
    /**
     * @generated from protobuf field: optional double bet = 1
     */
    bet?: number;
    /**
     * @generated from protobuf field: optional uint32 cheat = 2
     */
    cheat?: number;
    /**
     * @generated from protobuf field: optional uint32 itemIndex = 3
     */
    itemIndex?: number;
    /**
     * @generated from protobuf field: optional uint32 itemID = 4
     */
    itemID?: number;
    /**
     * @generated from protobuf field: optional int32 currencyNumber = 7
     */
    currencyNumber?: number;
    /**
     * @generated from protobuf field: optional double mallBet = 8
     */
    mallBet?: number;
    /**
     * @generated from protobuf field: optional uint32 vip = 9
     */
    vip?: number;
    /**
     * @generated from protobuf field: optional bytes customReq = 10
     */
    customReq?: Uint8Array;
    /**
     * @generated from protobuf field: optional bool isFree = 11
     */
    isFree?: boolean;
    /**
     * @generated from protobuf field: optional uint32 Extra = 12
     */
    extra?: number;
    /**
     * @generated from protobuf field: optional uint32 Level = 13
     */
    level?: number;
    /**
     * @generated from protobuf field: optional int32 BDType = 14
     */
    bDType?: number;
    /**
     * @generated from protobuf field: optional int32 BBTimes = 15
     */
    bBTimes?: number;
    /**
     * @generated from protobuf field: optional double AccuMallBet = 16
     */
    accuMallBet?: number;
    /**
     * @generated from protobuf field: optional double GameMallBet = 17
     */
    gameMallBet?: number;
    /**
     * @generated from protobuf field: optional int32 GameMallType = 18
     */
    gameMallType?: number;
    /**
     * @generated from protobuf field: optional bool isr = 19
     */
    isr?: boolean;
    /**
     * @generated from protobuf field: optional int32 BuyCustomType = 20
     */
    buyCustomType?: number;
    /**
     * @generated from protobuf field: optional bool isturbo = 21
     */
    isturbo?: boolean;
    /**
     * @generated from protobuf field: optional serverProto.ItemSpin item = 22
     */
    item?: ItemSpin;
    /**
     * @generated from protobuf field: optional serverProto.MallSpin mall = 23
     */
    mall?: MallSpin;
    /**
     * @generated from protobuf field: optional serverProto.BuySpin buy = 24
     */
    buy?: BuySpin;
    /**
     * @generated from protobuf field: optional serverProto.SpecialCase special = 25
     */
    special?: SpecialCase;
    /**
     * @generated from protobuf field: optional serverProto.ExtraSpin extraSpin = 26
     */
    extraSpin?: ExtraSpin;
    /**
     * @generated from protobuf field: optional serverProto.FreeSpin freeSpin = 27
     */
    freeSpin?: FreeSpin;
    /**
     * @generated from protobuf field: optional bytes clientRecord = 28
     */
    clientRecord?: Uint8Array;
    /**
     * @generated from protobuf field: optional serverProto.SpecialBonus specialBonus = 29
     */
    specialBonus?: SpecialBonus;
    /**
     * @generated from protobuf field: optional bool istournament = 30
     */
    istournament?: boolean;
}
// ...
export enum MallType {
    /**
     * @generated from protobuf enum value: NORMAL_MALL = 0;
     */
    NORMAL_MALL = 0,
    /**
     * @generated from protobuf enum value: GAME_MALL = 1;
     */
    GAME_MALL = 1
}
// ...
export const Request = new Request$Type();
// ...
export const freeSpinList = new freeSpinList$Type();
// ...
export const InfoReq = new InfoReq$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Browser$Type extends MessageType<Browser> {
    constructor() {
        super("serverProto.Browser", [
            { no: 1, name: "type", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "version", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "language", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "width", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
            { no: 5, name: "height", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
            { no: 6, name: "ratio", kind: "scalar", T: 2 /*ScalarType.FLOAT*/ }
        ]);
    }
    create(value?: PartialMessage<Browser>): Browser {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.type = "";
        message.version = "";
        message.language = "";
        message.width = 0;
        message.height = 0;
        message.ratio = 0;
        if (value !== undefined)
            reflectionMergePartial<Browser>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Browser): Browser {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string type */ 1:
                    message.type = reader.string();
                    break;
                case /* string version */ 2:
                    message.version = reader.string();
                    break;
                case /* string language */ 3:
                    message.language = reader.string();
                    break;
                case /* uint32 width */ 4:
                    message.width = reader.uint32();
                    break;
                case /* uint32 height */ 5:
                    message.height = reader.uint32();
                    break;
                case /* float ratio */ 6:
                    message.ratio = reader.float();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Browser, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string type = 1; */
        if (message.type !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.type);
        /* string version = 2; */
        if (message.version !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.version);
        /* string language = 3; */
        if (message.language !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.language);
        /* uint32 width = 4; */
        if (message.width !== 0)
            writer.tag(4, WireType.Varint).uint32(message.width);
        /* uint32 height = 5; */
        if (message.height !== 0)
            writer.tag(5, WireType.Varint).uint32(message.height);
        /* float ratio = 6; */
        if (message.ratio !== 0)
            writer.tag(6, WireType.Bit32).float(message.ratio);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message serverProto.Browser
 */
export const Browser = new Browser$Type();
// ...
export const GameInfoAck = new GameInfoAck$Type();
// ...
export const WalletInfo = new WalletInfo$Type();
// ...
export const PreferRoundShow = new PreferRoundShow$Type();
// ...
export const MallInfo = new MallInfo$Type();
// ...
export const GameMallInfo = new GameMallInfo$Type();
// ...
export const SpinReq = new SpinReq$Type();